dont serialize optional parameters

+++ Drover.Api/Services/IProjectService.cs
@@ -13,11 +13,11 @@ namespace Drover.Api.Services
 
         Task<List<Project>> GetActiveProjects(int? page, CancellationToken cancellationToken);
 
-        Task<Project> GetProject(int id, CancellationToken cancellationToken);
+        Task<Project> GetProject(long id, CancellationToken cancellationToken);
 
         Task CreateProject(string name, string websiteUrl, bool isActive, bool isPublic, bool guestsSeeGuests, CancellationToken cancellationToken);
 
-        Task AddMember(int projectId, int memberId, CancellationToken cancellationToken);
+        Task AddMember(long projectId, int memberId, CancellationToken cancellationToken);
 
         /// <summary>
         /// Invite a Guest using his E-Mail or add an existing via its Id
@@ -26,8 +26,8 @@ namespace Drover.Api.Services
         /// <param name="guestId"></param>
         /// <param name="email"></param>
         /// <returns></returns>
-        Task AddGuest(int projectId, int guestId, string email, CancellationToken cancellationToken);
+        Task AddGuest(long projectId, int guestId, string email, CancellationToken cancellationToken);
 
-        Task DeleteProject(int projectId, CancellationToken cancellationToken);
+        Task DeleteProject(long projectId, CancellationToken cancellationToken);
     }
 }
